book_management/pyproject.toml
[tool.poetry]
name = "book-management"
version = "0.1.0"
description = "도서 관리 앱"
authors = ["Seongeun Yu"]
license = "MIT License"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.95.0"
sqlalchemy = "^1.4.47"
uvicorn = "^0.21.1"
psycopg2 = "^2.9.5"
httpx = "^0.23.3"
dependency-injector = "^4.41.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
orjson = "^3.8.9"
asgi-lifespan = "^2.1.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
redis = "^4.5.4"
asyncpg = "^0.27.0"
polyfactory = "^1.0.0a0"

[tool.poetry.group.dev.dependencies]
ipython = "^8.11.0"
pytest = "^7.2.2"
faker = "^18.3.1"
pytest-asyncio = "^0.21.0"
pytest-factoryboy = "^2.5.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
